# api/docs/swagger.yaml
definitions:
...
  booking.BookingCoach:
    properties:
      access_status:
        type: string
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      payment:
        type: integer
      start_date:
        type: string
      subscription_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  booking.BookingGroup:
    properties:
      access_status:
        type: string
      count:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      payment:
        type: integer
      start_date:
        type: string
      subscription_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  booking.ListBookingCoachResponse:
    properties:
      booking_coach:
        items:
          $ref: '#/definitions/booking.BookingCoach'
        type: array
    type: object
  booking.ListBookingGroupResponse:
    properties:
      booking_group:
        items:
          $ref: '#/definitions/booking.BookingGroup'
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/booking-coach:
    get:
      consumes:
      - application/json
      description: Get a list of booking coach records.
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by subscription ID
        in: query
        name: subscription_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListBookingCoachResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Booking Coach
      tags:
      - BookingCoach
    post:
      consumes:
      - application/json
      description: Create a new booking coach with the provided details.
      parameters:
      - description: Booking Coach details
        in: body
        name: booking_coach
        required: true
        schema:
          $ref: '#/definitions/booking.BookingCoach'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.BookingCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Booking Coach
      tags:
      - BookingCoach
  /v1/booking-coach/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking coach by its ID.
      parameters:
      - description: Booking Coach ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Booking Coach
      tags:
      - BookingCoach
    get:
      consumes:
      - application/json
      description: Get details of a booking coach by its ID.
      parameters:
      - description: Booking Coach ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Booking Coach by ID
      tags:
      - BookingCoach
    put:
      consumes:
      - application/json
      description: Update an existing booking coach.
      parameters:
      - description: Booking Coach ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated booking coach
        in: body
        name: booking_coach
        required: true
        schema:
          $ref: '#/definitions/booking.BookingCoach'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingCoach'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Booking Coach
      tags:
      - BookingCoach
  /v1/booking-group:
    get:
      consumes:
      - application/json
      description: Get a list of booking group records.
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListBookingGroupResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Booking Group
      tags:
      - BookingGroup
    post:
      consumes:
      - application/json
      description: Create a new booking group with the provided details.
      parameters:
      - description: Booking Group details
        in: body
        name: booking_group
        required: true
        schema:
          $ref: '#/definitions/booking.BookingGroup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/booking.BookingGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Booking Group
      tags:
      - BookingGroup
  /v1/booking-group/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking group by its ID.
      parameters:
      - description: Booking Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Booking Group
      tags:
      - BookingGroup
    get:
      consumes:
      - application/json
      description: Get details of a booking group by its ID.
      parameters:
      - description: Booking Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Booking Group by ID
      tags:
      - BookingGroup
    put:
      consumes:
      - application/json
      description: Update an existing booking group.
      parameters:
      - description: Booking Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated booking group
        in: body
        name: booking_group
        required: true
        schema:
          $ref: '#/definitions/booking.BookingGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingGroup'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Booking Group
      tags:
      - BookingGroup
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"